+++ CHANGELOG.md
@@ -13,6 +13,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - Improved policy store API input validation. #528
 - Extended datamodel of EdcPolicyPermissionConstraint to include andConstraints
 - Marked createAccessPolicy requests with deprecation mark.
+- remove edc namespace from EdcContractAgreementNegotiationResponse and
+  EdcContractAgreementsResponse eclipse-tractusx/traceability-foss#963
 
 
 ## Added